修改返回值

+++ fileop.cpp
@@ -39,6 +39,7 @@ void FileOp::dirIterate(QString dir_path, void(*operation)(QString path, void *a
 
 void FileOp::zipFolder(const QString unziped_path)
 {
+   if(!QFile::exists(unziped_path)) return;
    QString zip_path = unziped_path.mid(0, unziped_path.lastIndexOf('.')) + ".docx";
    if(QFile::exists(zip_path)) QFile::remove(zip_path);
    zipper::Zipper *zipper = new zipper::Zipper(zip_path.toStdString(), zipper::Zipper::openFlags::Overwrite);
@@ -51,7 +52,7 @@ void FileOp::zipFolder(const QString unziped_path)
 
 QString FileOp::unzipFolder(QString ziped_path)
 {
-   if(!QFile::exists(ziped_path)) return;
+   if(!QFile::exists(ziped_path)) return "";
    int dot_pos = ziped_path.lastIndexOf('.');
    QString dir_path = ziped_path.mid(0, dot_pos) + ".cache";
    zipper::Unzipper *unzipper = new zipper::Unzipper(ziped_path.toStdString());
